extract essentials factory
