+++ core/src/main/java/org/jdbi/v3/core/mapper/ColumnMappers.java
@@ -38,6 +38,7 @@ public class ColumnMappers implements JdbiConfig<ColumnMappers> {
         factories.add(new OptionalMapperFactory());
         factories.add(new PrimitiveMapperFactory());
         factories.add(new BoxedMapperFactory());
+        factories.add(new EssentialsMapperFactory());
         factories.add(new BuiltInMapperFactory());
         factories.add(new SqlArrayMapperFactory());
     }
